Refactored structs into models module
